Attempt 8 mobile scrolling fix
I'm losing my mind. Mobile is now responsive, which means that the IF statement was always evaluating to TRUE. Which is wack because the first time was with !== INPUT and the second time was ==. Two contradictory terms that both evaluated to TRUE. So I'm gonna try and print out what the nodeName is for mobile.